+++ app/src/main/java/ru/kostenko/nework/dto/User.kt
@@ -1,7 +1,7 @@
 package ru.kostenko.nework.dto
 
 data class User(
-    override val id: Long = 0,
+    override val id: Int = 0,
     val login: String = "",
     val name: String = "",
     val avatar: String? = null
